Clarifying trust transfer
It is just a single change, but I think the current wording "SHOULD NOT assume" is not what you want to say. In fact, it should be written "SHALL  NOT assume". The new wording still allows the transfer of trust but makes clear, that nobody has a right to expect, that trust can be transferred.